requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langchain-text-splitters
chromadb
python-dotenv
tiktoken
